+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update ESLint configuration and refactoring (#52)
 - Add drag-and-drop for rows and fields (#54)
 - Update custom code when switching the value editor from manual to JS (#59)
+- Add Variable Editor with Fields and Java Script (#58)
 
 ## 3.0.0 (2023-06-28)